Fix changes. design pages to be more responsive
