+++ FromLocalsToLocals/Utilities/ExtensionMethods.cs
@@ -1,5 +1,7 @@
 ﻿using FromLocalsToLocals.Models;
 using Microsoft.AspNetCore.Http;
+using SendGrid;
+using SendGrid.Helpers.Mail;
 using SuppLocals;
 using System;
 using System.IO;
@@ -54,5 +56,20 @@ namespace FromLocalsToLocals.Utilities
             return null;
         }
 
+        public static async Task ExceptionSender(this Exception e)
+        {
+            var key = Config.Send_Grid_Key;
+            var client = new SendGridClient(key);
+
+            var from = new EmailAddress("fromlocalstolocals@gmail.com", "Exception message");
+            var subject = "Exception receiver";
+            var to = new EmailAddress("lukasstc223@gmail.com", "Dear User");
+            var plainTextContent = "We have this exception:";
+
+
+            var htmlContent = "<!DOCTYPE html><html><head><meta charset=\"UTF-8\"></head><body> " + e.ToString();
+            var msg = MailHelper.CreateSingleEmail(from, to, subject, plainTextContent, htmlContent);
+            var response = await client.SendEmailAsync(msg);
+        }
     }
 }
