+++ FromLocalsToLocals/Views/Home/Index.cshtml
@@ -19,6 +19,8 @@
             box-shadow: 0 0 15px rgba(10, 10, 10, 0.3);
         }
     </style>
+
+    <link href="https://unpkg.com/aos@2.3.1/dist/aos.css" rel="stylesheet">
 </head>
 <body>
     <div class="row" style="background: url(/Assets/vendorM.png);">
@@ -39,7 +41,7 @@
     <p id="mapTitle" class="mt-3"></p>
     <h4 class="mt-3 mb-3" style="text-align: center; font-weight: bold">@_localizer["Discover all nearby vendors with just a few clicks."]</h4>
     <div class="container">
-        <div class="row mb-3 justify-content-center d-flex">
+        <div class="row mb-3 justify-content-center d-flex" data-aos="zoom-in">
             @Html.Partial("_Map", Model.AllVendors)
         </div>
     </div>
@@ -59,7 +61,7 @@
             @foreach (var item in Model.PopularVendors)
             {
                 <div class="col-sm-@(12 / itemsInRow) mb-3 mt-3 d-flex justify-content-center">
-                    <div class="card" style="width:inherit; height:inherit; background-color: #F7F7F7;">
+                    <div class="card" data-aos="fade-up"  style="width:inherit; height:inherit; background-color: #F7F7F7;">
                         @if (item.Image == null)
                         {
                             <img src="~/Assets/localSeller.png" class="mt-3" style="width:160px ; height:150px; border-radius: 50%; align-self:center;" />
@@ -182,3 +184,7 @@
 </body>
 
 <script src="~/js/following.js"></script>
+<script src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
+<script>
+    AOS.init();
+</script>
