+++ FromLocalsToLocals/Views/Shared/Components/NewsFeed/Default.cshtml
@@ -7,13 +7,13 @@
 <div class="container">
     <div class="row">
         <div class="col-md-12">
-            <div id="content1" onscroll="scrollFeedDetails()" class="content content-full-width newsFeedPanel">
+            <div id="content1" onscroll="scrollFeedDetails()" class="content content-full-width newsFeedPanel"  >
                 <!-- begin profile-content -->
                 <div class="profile-content">
                     <!-- begin tab-content -->
                     <div class="tab-content p-0">
                         <!-- begin #profile-post tab -->
-                        <div class="tab-pane fade active show" id="profile-post">
+                        <div class="tab-pane fade active show" id="profile-post" data-aos="fade-left" >
                             <!-- begin timeline -->
                             <ul class="timeline" id="postsUL">
