+++ FromLocalsToLocals/Views/Shared/Error.cshtml
@@ -4,27 +4,88 @@
 @{
     ViewData["Title"] = "Error";
 }
+<style>
 
-<html lang="en" dir="auto">
+    /*======================
+    404 page
+=======================*/
+
+    .page_404 {
+        padding: 40px 0;
+        background: #fff;
+        font-family: 'Arvo', serif;
+    }
+
+        .page_404 img {
+            width: 100%;
+        }
+
+    .four_zero_four_bg {
+        background-image: url(https://cdn.dribbble.com/users/285475/screenshots/2083086/dribbble_1.gif);
+        height: 400px;
+        background-position: center;
+    }
 
 
-<head>
-    <meta charset="utf-8" />
-    <link href="~/css/error.css" rel="stylesheet" type="text/css">
-</head>
+        .four_zero_four_bg h1 {
+            font-size: 80px;
+        }
+
+        .four_zero_four_bg h3 {
+            font-size: 80px;
+        }
+
+    .link_404 {
+        color: #fff !important;
+        padding: 10px 20px;
+        background: #39ac31;
+        margin: 20px 0;
+        display: inline-block;
+    }
+
+    .contant_box_404 {
+        margin-top: -50px;
+    }
+
+    footer {
+        display: none;
+    }
+
+
+    header {
+        display: none;
+    }
+
+    body{
+        text-align: center;
+    }
+</style>
 
 <body>
-    <div class="error-page">
-        <div class="content">
-            <h1 data-text="404">404</h1>
-            <h4 data-text="Oops! Something went wrong">Opps! Something went wrong</h4>
-            <p>Sorry, the current page you are looking doesn't work properly!</p>
-            <div class="btns">
-                <a asp-controller="Home" asp-action="Index">return home</a>
-                <a asp-controller="Home" asp-action="ReportBug">report problem</a>
+    <section class="page_404">
+        <div class="container-fluid">
+            <div class="row">
+                <div class="col-sm-12 ">
+                    <div class="col-sm-10 col-sm-offset-1  text-center">
+                        <div class="four_zero_four_bg">
+                            <h1 class="text-center ">404</h1>
+
+                        </div>
+
+                        <div class="contant_box_404">
+                            <h3 class="h2">
+                                Look like you're lost
+                            </h3>
+
+                            <p>the page you are looking for not avaible!</p>
+
+                            <a href="" class="link_404">Go to Home</a>
+                        </div>
+                    </div>
+                </div>
             </div>
         </div>
-    </div>
-</body>
+    </section>
 
-</html>
+    </body>
+
