+++ FromLocalsToLocals/Views/Vendors/Details.cshtml
@@ -8,7 +8,7 @@
     Layout = "~/Views/Shared/_Layout.cshtml";
 
     var userForFeed = await userManager.GetUserAsync(User);
-    if (userForFeed !=null && userForFeed.Id != Model.User.Id) userForFeed = null;
+    if (userForFeed != null && userForFeed.Id != Model.User.Id) userForFeed = null;
 
     var timeSpan = new TimeSpan(0);
